fix: 14 correctness and safety bugs across grading, picks, and UI
- gradeEngine: break→continue so remaining picks aren't skipped when a
  matched game has no line; NaN guard on parseFloat(_rawLine) for "pk"/""
- picks/route: wrap JSON.parse(cachedOdds) in try/catch; check parlay
  rollback delete error; allow null leg odds from screenshots (validation
  + legRows mapping); update server-side leg odds validation to skip null
- sports/route: lower teamNamesMatch thresholds (>3→>=3, >4→>=3) to
  fix Red Sox / Cubs / White Sox matching
- HistoryTab: null-coerce parlay leg odds to null not 0; optional-chain
  live?.liveStatus to prevent null dereference
- contestValidation: parseInt→parseFloat for fractional odds like 150.5
- admin/route: destructure and check error from XP profile.update
- chat/route: check Supabase errors on all ban/mute/unmute operations
- Sidebar: replace 500ms setInterval polling with storage event listener
- pregenerate-analysis: collapse dead fetchPlayerProps body (~94 lines)
  to a single early-return stub; dead code was after return ''
- odds.js: replace dead totals parsing (API never returns totals market)
  with null constants; keep return shape for compatibility
- cron/grade-picks: add console.warn to two silent catch blocks

+++ src/lib/contestValidation.js
@@ -54,7 +54,7 @@ export function validateContestEntry(pick, rules = {}) {
   }
 
   // ── Odds range ──────────────────────────────────────────────────────────────
-  const odds = parseInt(pick.odds);
+  const odds = parseFloat(pick.odds);
   if (rules.require_odds && (!pick.odds || isNaN(odds))) {
     errors.push('Odds are required for contest entries.');
   }
